Hack for dygraphs library
Dygraphs is cool,
but for Magnitude graph i need reverted yrange plot. This kind of plots don't work properly on zooming events. So i've decided to multipy Mag by -1 and format label tics. To prevent -Mag in the legend i also need to multiply it by -1. This commit did the later.

previous commit was made to see the diff made for hack

Signed-off-by: Dmitriy Blinov <dmitriy.blinov@gmail.com>